requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.6.0

# LLM & AI
langchain-openai>=0.1.0
langgraph>=0.1.0
firecrawl>=1.0.0

# Streamlit interface
streamlit>=1.25.0
pandas>=2.1.0
pillow>=10.0.0

# Optional (for advanced UI features)
numpy>=1.26.0
requests>=2.31.0
